pyproject.toml
[project]
name = "chineseselflearningsystem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "mysql-connector-python>=9.6.0",
    "groq>=0.4.2",
    "bcrypt>=4.0.1",
    "python-dotenv>=1.0.0",
]
